space bitches can kill

+++ src/systems/particle_system.rb
@@ -4,8 +4,10 @@ class ParticleSystem < System
       particle.spatial_state.thrust(particle.speed)
       particle.color.a -= 0.05
 
-      particle.lifetime.update
-      particle.destroy if particle.lifetime.dead?
+      if particle.lifetime
+        particle.lifetime.update
+        particle.destroy if particle.lifetime.dead?
+      end
     end
   end
 end
